recommendations-api/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pymongo==4.11.1
python-dotenv==1.0.1
numpy==2.2.3
anthropic==0.39.0
sse-starlette==2.1.3
